Move MsgContext to interfaces dir
