+++ src/bot/command.ts
@@ -1,4 +1,4 @@
-import {MsgContext} from './msg_context';
+import {MsgContext} from '../interfaces';
 
 export interface Command {
   matches(name: string): boolean;
