+++ src/bot/command_handler.ts
@@ -1,5 +1,5 @@
 import {Command} from './command';
-import {MsgContext} from './msg_context';
+import {MsgContext} from '../interfaces';
 
 export class CommandHandler {
   constructor(private readonly commands: Command[]) {
