+++ src/bot/commands/start.ts
@@ -1,7 +1,7 @@
 import {EntityManager} from 'typeorm';
 import {Command} from '../command';
 import {Session} from '../../models';
-import {MsgContext} from '../msg_context';
+import {MsgContext} from '../../interfaces';
 
 export class StartCommand implements Command {
   constructor(private readonly manager: EntityManager) {
