+++ test/fakes/msg_context.ts
@@ -1,5 +1,5 @@
 import {TextChannel} from 'discord.js';
-import {MsgContext} from '../../src/bot/msg_context';
+import {MsgContext} from '../../src/interfaces';
 import {Session} from '../../src/models/session';
 import {FakeTextChannel} from './text_channel';
